refactor(js-debugger): Extract inspector helpers from integration test
Moved CDP/inspector helper functions to separate utility file to fix
max-lines ESLint violation (417 → 185 lines).

Extracted functions:
- launchInspector
- createCdpCommandSender
- collectConsoleEvents
- openWebSocket
- parseInspectorUrl
- terminateProcess

All tests continue to pass. File now well under 400 line limit.
